Floats can now be read

+++ Coursework 2/src/decodeCommands.h
@@ -5,7 +5,7 @@
 #include "messages.h"
 
 extern volatile uint64_t* key;
-extern volatile int32_t tar_velocity;
+extern volatile float tar_velocity;
 extern volatile int32_t tar_rotations;
 extern volatile uint32_t motorPWM;
 extern volatile uint32_t err;
